requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
SQLAlchemy==2.0.38
psycopg2-binary==2.9.10
alembic==1.14.1
pydantic-settings==2.8.0
pytest==8.3.4
httpx==0.28.1
